Pull request #804: AG-51940 Add executor and requests for platform API
Merge in SCITER/adguard-mini from AG-51940 to master

Squashed commit of the following:

commit 34a7a0601bbef02152750fe6b9e0f556ef7fde4e
Merge: a463f3fb ecbfbf54
Author: Vladislav Abdulmyanov <v.abdulmyanov@adguard.com>
Date:   Thu Apr 30 14:23:11 2026 +0300

    Merge branch 'master' into AG-51940

commit a463f3fb37d8e4623207f32ca0957343fd6acbd0
Author: Pavel Kizin <p.kizin@adguard.com>
Date:   Fri Apr 24 14:16:44 2026 +0300

    AG-51940 Fix build

commit 72cf74ad5b8158ed0349f805f9a2650fc7648020
Merge: b8823d93 e63e3b0c
Author: Pavel Kizin <p.kizin@adguard.com>
Date:   Fri Apr 24 12:32:31 2026 +0300

    Merge remote-tracking branch 'remotes/origin/master' into AG-51940

commit b8823d9334b0d1c21994600b6991e891e6d55232
Author: Pavel Kizin <p.kizin@adguard.com>
Date:   Thu Apr 23 15:20:08 2026 +0300

    AG-51940 Fix import

commit b9fb607c91cbbbc9ee6c0b85883c22fb0b7123cd
Author: Pavel Kizin <p.kizin@adguard.com>
Date:   Thu Apr 23 15:08:38 2026 +0300

    AG-51940 Fix review

commit c6c8d5015197eca8d271020074c70781a8ac40c5
Author: Pavel Kizin <p.kizin@adguard.com>
Date:   Thu Apr 23 13:44:22 2026 +0300

    AG-51940 Fix review

commit 93e78d01e7a7383643d2bf943a00cb97279065ee
Author: Pavel Kizin <p.kizin@adguard.com>
Date:   Thu Apr 23 13:41:48 2026 +0300

    AG-51940 Update dependencies

commit 31de453d3dcc41eb13b3036a8ce7be1dd934dc31
Merge: 60e2b15b be4cbf99
Author: Pavel Kizin <p.kizin@adguard.com>
Date:   Mon Apr 20 18:35:24 2026 +0300

    Merge remote-tracking branch 'remotes/origin/master' into AG-51940

commit 60e2b15bdd5dc6a1e862a4d8ca6a24fcac932520
Author: Pavel Kizin <p.kizin@adguard.com>
Date:   Mon Apr 20 18:21:53 2026 +0300

    AG-51940 Fix review

commit 107d7af43bac5bdad8a479d181df86ba68d029e3
Author: Pavel Kizin <p.kizin@adguard.com>
Date:   Mon Apr 20 14:51:11 2026 +0300

    AG-51940 Fix review

commit df75afb7136994be9e8209508d37264a1cdfd313
Author: Pavel Kizin <p.kizin@adguard.com>
Date:   Mon Apr 20 14:28:44 2026 +0300

    AG-51490 Update REUSE.toml

commit 38aa26fe45aadbf51b018436be0843dba5c17586
Author: Pavel Kizin <p.kizin@adguard.com>
Date:   Fri Apr 17 16:56:43 2026 +0300

    AG-51940 Update dependencies

commit b7e279a2410cc6e918896686eb9fd9e89a90eda3
Author: Pavel Kizin <p.kizin@adguard.com>
Date:   Fri Apr 17 16:55:28 2026 +0300

    AG-51940 Add executor and requests for platform API

+++ AdguardMini/Scripts/updateProtoSchema.sh
@@ -60,9 +60,15 @@ echo "Cleanup and make dirs"
 echo "=================================================================="
 echo
 
-find "$TYPESCRIPT_SCHEMA_OUTPUT_DIR/callbacks" -type f -name "*.ts" -delete
-find "$TYPESCRIPT_SCHEMA_OUTPUT_DIR/services" -type f -name "*.ts" -delete
-find "$TYPESCRIPT_SCHEMA_OUTPUT_DIR/types" -type f -name "*.ts" -delete
+if [ -d "$TYPESCRIPT_SCHEMA_OUTPUT_DIR/callbacks" ]; then
+    find "$TYPESCRIPT_SCHEMA_OUTPUT_DIR/callbacks" -type f -name "*.ts" -delete
+fi
+if [ -d "$TYPESCRIPT_SCHEMA_OUTPUT_DIR/requests" ]; then
+    find "$TYPESCRIPT_SCHEMA_OUTPUT_DIR/requests" -type f -name "*.ts" -delete
+fi
+if [ -d "$TYPESCRIPT_SCHEMA_OUTPUT_DIR/types" ]; then
+    find "$TYPESCRIPT_SCHEMA_OUTPUT_DIR/types" -type f -name "*.ts" -delete
+fi
 find "$SWIFT_SCHEMA_OUTPUT_DIR" -type f -name "*.swift" -delete
 
 echo "Done!"
